+++ TwinPushSDKDemo/RichNotificationService/NotificationService.h
@@ -1,0 +1,13 @@
+//
+//  NotificationService.h
+//  RichNotificationService
+//
+//  Created by Guillermo Gutiérrez Doral on 17/4/17.
+//  Copyright © 2017 TwinCoders. All rights reserved.
+//
+
+#import <UserNotifications/UserNotifications.h>
+
+@interface NotificationService : UNNotificationServiceExtension
+
+@end